requirements.txt
aniso8601==3.0.2
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
backports.csv==1.0.7
beautifulsoup4==4.6.3
blinker==1.4
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
braintree==3.52.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
dataclasses==0.6
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-stubs==0.11.1
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
docutils==0.14
et-xmlfile==1.0.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
jmespath==0.9.4
lxml==4.3.3
Markdown==3.1
markdown2==2.3.7
MarkupSafe==1.1.1
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==6.0.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
oauthlib==3.0.1
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.2
parsel==1.5.1
Pillow==5.4.1
pluggy==0.9.0
promise==2.2.1
psutil==5.4.8
psycopg2==2.8.1
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.6.8
pytest==4.3.1
pytest-runner==4.4
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
Rx==1.6.1
s3transfer==0.2.0
Scrapy==1.6.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
stripe==2.24.1
tablib==0.13.0
Twisted==18.9.0
typed-ast==1.3.1
typing==3.6.4
typing-extensions==3.7.2
unicodecsv==0.14.1
urllib3==1.24.1
w3lib==1.20.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
